Generate settings classes from a json file
