+++ src/GitHub.Exports/Settings/IPackageSettings.cs
@@ -1,0 +1,9 @@
+﻿// This is an automatically generated file, based on settings.json and IPackageSettings.tt
+namespace GitHub.Settings
+{
+    public interface IPackageSettings
+    {
+        void Save();
+        bool CollectMetrics { get; set; }
+    }
+}
