+++ src/GitHub.Exports/Settings/IPackageSettings.tt
@@ -1,0 +1,28 @@
+﻿<#@ template debug="false" hostspecific="true" language="C#" #>
+<#@ assembly name="System.Core" #>
+<#@ import namespace="System.Linq" #>
+<#@ import namespace="System.Text" #>
+<#@ import namespace="System.Collections.Generic" #>
+<#@ import namespace="System.IO" #>
+<#@ assembly name="$(SolutionDir)\build\$(Configuration)\Newtonsoft.Json.Dll" #>
+<#@ import namespace="Newtonsoft.Json.Linq" #>
+<#@ output extension=".cs" #>
+<#
+var file = this.Host.ResolvePath(@"..\..\common\settings.json");
+var json = JObject.Parse(File.ReadAllText(file));
+#>
+// This is an automatically generated file, based on settings.json and IPackageSettings.tt
+namespace GitHub.Settings
+{
+    public interface IPackageSettings
+    {
+        void Save();
+<#
+foreach (var j in json.Children().Children().Children()) {
+#>
+        <#= j["type"] #> <#= j["name"] #> { get; set; }
+<#
+}
+#>
+    }
+}
