+++ src/GitHub.VisualStudio/Settings/PackageSettingsGen.tt
@@ -1,0 +1,52 @@
+﻿<#@ template debug="false" hostspecific="true" language="C#" #>
+<#@ assembly name="System.Core" #>
+<#@ import namespace="System.Linq" #>
+<#@ import namespace="System.Collections.Generic" #>
+<#@ import namespace="System.IO" #>
+<#@ assembly name="$(TargetDir)\Newtonsoft.Json.Dll" #>
+<#@ import namespace="Newtonsoft.Json.Linq" #>
+<#@ output extension=".cs" #>
+<#
+var file = this.Host.ResolvePath(@"..\..\common\settings.json");
+var json = JObject.Parse(File.ReadAllText(file));
+#>
+// This is an automatically generated file, based on settings.json and PackageSettingsGen.tt
+using GitHub.Settings;
+
+namespace GitHub.VisualStudio.Settings {
+
+    public partial class PackageSettings : IPackageSettings
+    {
+
+<#
+foreach (var j in json.Children().Children().Children()) {
+#>
+        public <#= j["type"] #> <#= j["name"] #> { get; set; }
+<#
+}
+#>
+
+		void LoadSettings()
+		{
+<#
+foreach (var j in json.Children().Children().Children()) {
+#>
+			<#= j["name"] #> = (<#= j["type"] #>)settingsStore.Read("<#= j["name"] #>", <#= j["default"] #>);
+<#
+}
+#>
+		}
+
+		void SaveSettings()
+		{
+<#
+foreach (var j in json.Children().Children().Children()) {
+#>
+			settingsStore.Write("<#= j["name"] #>", <#= j["name"] #>);
+<#
+}
+#>
+		}
+
+	}
+}
