+++ src/common/settings.json
@@ -1,0 +1,9 @@
+﻿{
+	"settings": [
+		{
+			"name": "CollectMetrics",
+			"type": "bool",
+			"default": "true"
+		}
+	]
+}
